requirements.txt
ipython==5.0
six
matplotlib
jupyter
Pillow
numpy
absl-py
scipy
scikit-learn
bleach==1.5.0
tensorflow-gpu==1.8.0
psutil
tmuxp
requests
googledrivedownloader
tqdm
pip